roughly finish refactoring
