+++ src/App.tsx
@@ -1,11 +1,5 @@
-import DashboardPage from "pages/platform/dashboard/dashboardPage"
-import LandingPage from "pages/landing-page/HomePage"
-import Login from "pages/landing-page/login-and-signup/login-page/Login"
-import Signedup from "pages/landing-page/login-and-signup/signup-page/Signedup"
-import Signup from "pages/landing-page/login-and-signup/signup-page/Signup"
-import OnlinePractice from "pages/platform/practicing/PracticeHomepage/OnlinePractice"
-import RAList from "pages/platform/practicing/QuestionType/RA/QuestionList/RAList"
-import React from "react"
+import LandingPage from "pages/landing-page/LandingPage"
+import Platform from "pages/platform/Platform"
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom"
 import { ThemeProvider } from "styled-components"
 import { createGlobalStyle } from "styled-components"
@@ -26,13 +20,8 @@ function App() {
       <Router>
         <GlobalStyles />
         <Routes>
-          <Route path="/" element={<LandingPage />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/signup" element={<Signup />} />
-          <Route path="/signedup" element={<Signedup />} />
-          <Route path="/dashboard" element={<DashboardPage />} />
-          <Route path="/practice" element={<OnlinePractice />} />
-          <Route path="/practice/Speaking/RA" element={<RAList />} />
+          <Route path="/*" element={<LandingPage />} />
+          <Route path="/platform/*" element={<Platform />} />
         </Routes>
       </Router>
     </ThemeProvider>
