+++ src/pages/landing-page/homepage/HomePage.tsx
@@ -1,6 +1,3 @@
-import Footer from "pages/public-component/Footer/Footer"
-import Navbar from "pages/public-component/Navbar"
-
 import Hero from "./components/Hero/Hero"
 import { PracticeArea } from "./components/PracticeArea/PracticeArea"
 import StudyMaterial from "./components/StudyMaterial/StudyMaterial"
@@ -9,12 +6,10 @@ import Testimonial from "./components/Testimonial/Testimonial"
 const LandingPage = () => {
   return (
     <>
-      <Navbar />
       <Hero />
       <PracticeArea />
       <StudyMaterial />
       <Testimonial />
-      <Footer />
     </>
   )
 }
