+++ src/pages/landing-page/homepage/components/PracticeArea/PracticeArea.tsx
@@ -1,5 +1,5 @@
-import { PracticeAreaCards } from "pages/landing-page/components/PracticeArea/Components/PracticeAreaCard/PracticeAreaCard"
-import { WeeklyPredictionCard } from "pages/landing-page/components/PracticeArea/Components/WeeklyPredictionCard/WeeklyPredictionCard"
+import { PracticeAreaCards } from "pages/landing-page/homepage/components/PracticeArea/Components/PracticeAreaCard/PracticeAreaCard"
+import { WeeklyPredictionCard } from "pages/landing-page/homepage/components/PracticeArea/Components/WeeklyPredictionCard/WeeklyPredictionCard"
 import styled from "styled-components"
 
 const PracticeAreaWrapper = styled.div`
