+++ src/pages/landing-page/homepage/components/StudyMaterial/StudyMaterial.tsx
@@ -1,4 +1,4 @@
-import Card from "pages/landing-page/components/StudyMaterial/Card/Card"
+import Card from "pages/landing-page/homepage/components/StudyMaterial/Card/Card"
 import React from "react"
 import styled from "styled-components"
